refactor(i18n): take advantage of DopplerIntlProvider.double-with-ids-as-values in our tests
Also:
* Move flattenMessages inside i18n folder (because now, it does not have sense using it outside)
