+++ src/i18n/DopplerIntlProvider.double-with-ids-as-values.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { addLocaleData, IntlProvider } from 'react-intl';
 import en from 'react-intl/locale-data/en';
 import messages_en from './en.json';
-import { flattenMessages } from '../utils';
+import { flattenMessages } from './utils';
 
 addLocaleData([...en]);
