+++ src/utils.js
@@ -1,18 +1,3 @@
 export function timeout(ms) {
   return new Promise((resolve) => setTimeout(resolve, ms));
 }
-
-export function flattenMessages(nestedMessages, prefix = '') {
-  return Object.keys(nestedMessages).reduce((messages, key) => {
-    let value = nestedMessages[key];
-    let prefixedKey = prefix ? `${prefix}.${key}` : key;
-
-    if (typeof value === 'string') {
-      messages[prefixedKey] = value;
-    } else {
-      Object.assign(messages, flattenMessages(value, prefixedKey));
-    }
-
-    return messages;
-  }, {});
-}
